+++ CHANGELOG.md
@@ -24,6 +24,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Find TMX files in `Generated` and `User Modified` directories
 
+- Don't ask Bing translator to translate language names: https://github.com/sillsdev/l10nsharp/issues/66.
+  Also don't display the name a second time in parentheses if English and native name are identical.
+
 ## [4.0.0] - 2019-05-16
 
 ### Changed